use the new swiftui app lifecycle setup
